vinyl: don't take index ref in vy_cursor_new and vy_index_send
We take index->refs in these functions in order to ensure the index
won't be dropped while we're working with it in the tx thread. Actually,
this logic isn't specific to a particular backend and should be
implemented at the upper level. So let's drop it for now - this will
allow us to simplify vy_index reference counting, which we do need to
avoid use-after-free in workers.