vinyl: rework vy_index reference counting
A vy_index might be dropped while there are still tasks pending for it.
To avoid use-after-free we currently use a rather weird technique: the
drop operation decrements vy_index->refs, but doesn't schedule index
destruction - instead the dead index is added to the "shutdown" list and
the scheduler itself keeps polling it queueing indexes whose ref counter
dropped to zero for destruction.

Let's rework it as follows. As before, vy_index has a ref counter (now
called task_refs), which is taken by any task as well as by the tx
thread. However, now it is decremented only from worker threads upon
task completion and from VY_TASK_DROP handler, which is now scheduled
immediately on index drop operation. When task_refs drops to zero, the
index gets destroyed. Such a design allows us to remove the loop
iterating over all dead indexes from the scheduler. Actually, we could
drop the "shutdown" list and schedule VY_TASK_DROP work directly from
vy_index_drop(), but I left it to keep the patch simple (removing it
requires reworking scheduler stop operation) - we can do it later in a
separate patch.

Note, we don't need to protect task_refs with a mutex or make it atomic
as it is only inc/dec'ed when the task is queued to input/output
scheduler queues, where we hold scheduler->mutex, so I just use the
existing mutex to protect it.